implement advanced page
- bump littlefs version
- fix compiler warnings (c23 extensions, casting)
- add set_config_full api command to allow for full config backups/restores
- consolidate api header files
- bump pnpm packages
- fix eslint
- standardize some www styling (like button colors, etc.)
- implement advanced page
  - add config backup/restore functions, reboot, revert setup, and about/license info
- update api mock for new commands
- update config database for new config values

+++ src/sys/api/cmds/TEST/test_pwm.c
@@ -16,14 +16,14 @@
 #include "sys/configuration.h"
 #include "sys/print.h"
 
-#include "test_pwm.h"
+#include "test.h"
 
 /**
  * @return a pseudo-randomly generated number between 0 and 180
  */
 static f32 rand_180() {
     srand((u32)time_us());
-    return (f32)rand() / RAND_MAX * 180;
+    return (f32)rand() / (f32)RAND_MAX * 180.f;
 }
 
 /**
